Method types was changed to "Result" type.
In bussines layer, Bussiness methods returns messages or true.false infos when they is run. So that, I am controlling that returned results was successful or unsuccessfull. If the method is including an data or data list, its type is "IDataResult". If the method doesn't return any data, its type is just "IResult".