Result classes created.
In bussines layer, Bussiness methods returns messages or true,false information when they is run. So that, I am controlling that returned result was successful or  unsuccessfull. If the method is including an data or list of data, its type is "IDataResult". If the method doesn't return any data, its type is just "IResult".